@@ -1,6 +1,6 @@
-import {HttpClientModule, provideHttpClient, withInterceptors, withJsonpSupport} from '@angular/common/http';
-import {ApplicationConfig, importProvidersFrom} from '@angular/core';
-import { provideRouter } from "@angular/router";
+import {provideHttpClient, withJsonpSupport} from '@angular/common/http';
+import {ApplicationConfig} from '@angular/core';
+import {provideRouter} from "@angular/router";
 import routerConfig from "./routes";
 import {httpInterceptorProviders} from "../../services/http.interceptor";
 
@@ -8,7 +8,6 @@ export const appConfig: ApplicationConfig = {
   providers: [
     provideHttpClient(withJsonpSupport()),
     provideRouter(routerConfig),
-    importProvidersFrom(HttpClientModule),
     httpInterceptorProviders
   ]
 };
